Fix IMC object allocation counts
In these cases, the object's class isn't actually `Object`, but the
singleton class for that instance.  The inline method cache retains a
reference to the singleton class, and the singleton class retains a
reference to the object from which it was created.

In these tests, the inline method cache hasn't been deleted, so the last
object that was referenced will be retained.  In other words, the
minimum number of live objects when executing this code will always be
1.  We can call `run` multiple times, then GC, then prove that only one
object is retained even though we made a few objects.