Attempt to be smarter about flushing
Instead of picking a magic number and we can create an object in memory that is not retained, we then trace the object and while we detect that it is in memory we repeatedly flush the stack and create new objects. Once the object is out of memory then we can determine that it’s been removed by GC. 

The theory here is that since the object we are tracing is more recent than the object that the user is tracing it will be more likely that GC will have cleared any unused objects before getting to the one we just created.

It’s still not perfect, but hopefully it’s better than it was previously.